requirements.txt
google-cloud-storage
google-generativeai
pypdf
fastapi
uvicorn
google-adk
zeep
requests
python-dotenv
google-cloud-aiplatform 
websockets
google-cloud-discoveryengine
pymysql